Switches to new quality settings in case the older ones error out. This is an ugly hack. An ugly ugly hack.

+++ twitchy.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 # Requires: python3, livestreamer, requests
-# rev = 51
+# rev = 52
 
 
 import sys
@@ -188,6 +188,14 @@ class Options:
 
 		""" Run time option """
 		conky_run = False
+
+		""" Map to alternate quality settings """
+		alternate_quality = {
+			'low': '360p',
+			'medium': '480p',
+			'high': '720p',
+			'source': 'best'}
+
 	except:
 		print(Colors.RED + ' Error getting options. Running --configure:' + Colors.ENDC)
 		configure_options('TheDudeAbides')
@@ -729,6 +737,8 @@ class Playtime:
 	def __init__(self, final_selection, stream_quality, display_name, game_name, show_chat, channel_name_if_vod):
 		self.final_selection = final_selection
 		self.stream_quality = stream_quality
+		self.stream_quality_alternate = Options.alternate_quality[self.stream_quality]
+		self.alternate_quality_tried = False  # In case the alternate quality doesn't work either
 		self.display_name = display_name
 		self.game_name = game_name
 		self.show_chat = show_chat
@@ -753,19 +763,34 @@ class Playtime:
 		if self.show_chat is True:
 			try:
 				webbrowser.get('chromium').open_new('--app=http://www.twitch.tv/%s/chat?popout=' % self.final_selection)
-			except:
+			except webbrowser.Error:
 				webbrowser.open_new('http://www.twitch.tv/%s/chat?popout=' % self.final_selection)
 
 		if self.channel_name_if_vod is None:
 			print(' ' + Colors.WHITE + self.display_name + Colors.ENDC + ' | ' + Colors.WHITE + self.stream_quality.title() + Colors.ENDC)
 			player_final = Options.player_final + ' --title ' + self.display_name.replace(' ', '')
-			self.args_to_subprocess = "{3} twitch.tv/'{0}' '{1}' --player '{2}' --hls-segment-threads 3 --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(self.final_selection, self.stream_quality, player_final, Options.backend)
+
+			self.args_to_subprocess = "{3} twitch.tv/'{0}' '{1}' --player '{2}' --hls-segment-threads 3 --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(
+				self.final_selection, self.stream_quality, player_final, Options.backend)
+
+			# Alternate quality in case the default one errors out
+			self.args_to_subprocess_alternate = "{3} twitch.tv/'{0}' '{1}' --player '{2}' --hls-segment-threads 3 --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(
+				self.final_selection, self.stream_quality_alternate, player_final, Options.backend)
 		else:
 			print(' ' + Colors.WHITE + self.display_name + ': ' + self.video_title_if_vod + Colors.ENDC + ' | ' + Colors.WHITE + self.stream_quality.title() + Colors.ENDC)
 			player_final = Options.player_final + ' --title ' + self.display_name
-			self.args_to_subprocess = "{3} '{0}' '{1}' --player '{2}' --hls-segment-threads 3 --player-passthrough=hls --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(self.final_selection, self.stream_quality, player_final, Options.backend)
+
+			self.args_to_subprocess = "{3} '{0}' '{1}' --player '{2}' --hls-segment-threads 3 --player-passthrough=hls --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(
+				self.final_selection, self.stream_quality, player_final, Options.backend)
+
+			# Alternate quality in case the default one errors out
+			self.args_to_subprocess_alternate = "{3} '{0}' '{1}' --player '{2}' --hls-segment-threads 3 --player-passthrough=hls --http-header Client-ID=guulhcvqo9djhuyhb2vi56wqnglc351".format(
+				self.final_selection, self.stream_quality_alternate, player_final, Options.backend)
 
 		self.args_to_subprocess = shlex.split(self.args_to_subprocess)
+		self.args_to_subprocess_alternate = shlex.split(self.args_to_subprocess_alternate)
+
+		# Starts with the default quality setting
 		self.livestreamer_process = subprocess.Popen(self.args_to_subprocess, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
 
 	def time_tracking(self):
@@ -853,13 +878,22 @@ def playtime_instances(final_selection):
 				if playtime_instance[k].livestreamer_process.returncode == 1:
 					stream_error = playtime_instance[k].livestreamer_process.stdout.read().decode('utf-8').split('\n')
 					error_message = [er for er in stream_error if 'error:' in er]
-					print(' ' + Colors.RED + playtime_instance[k].display_name + Colors.ENDC + ' (' + error_message[0] + ')')
-					# print(' ' + Colors.RED + playtime_instance[k].display_name + Colors.ENDC + ' (' + ' '.join(playtime_instance[k].args_to_subprocess) + ')')
-					database.execute("DELETE FROM miscellaneous WHERE Name = '{0}'".format(playtime_instance[k].display_name))
-					database.commit()
+
+					""" This hack is needed because the Twitch API does not give quality settings
+					We move on to the alternate quality settings specified in the Options class
+					OR we error out. It's a good day to die. """
+					if error_message[0][:30] == 'error: The specified stream(s)' and playtime_instance[k].alternate_quality_tried is False:
+						playtime_instance[k].livestreamer_process = subprocess.Popen(
+							playtime_instance[k].args_to_subprocess_alternate, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
+						playtime_instance[k].alternate_quality_tried = True
+					else:
+						print(' ' + Colors.RED + playtime_instance[k].display_name + Colors.ENDC + ' (' + error_message[0] + ')')
+						database.execute("DELETE FROM miscellaneous WHERE Name = '{0}'".format(playtime_instance[k].display_name))
+						database.commit()
+						playing_streams.remove(k)
 				else:
 					playtime_instance[k].time_tracking()
-				playing_streams.remove(k)
+					playing_streams.remove(k)
 		try:
 			keypress, o, e = select.select([sys.stdin], [], [], 0.8)
 			if keypress:
